PortPublicationsの切り出し
